// src/models/Appointment.ts
export class Appointment{
    idAppoint: number = 0;
    dateAppoint: string  = '';
    addTakeOffAppoint: boolean = false;
    stateAppoint: string = '';
    typeservice: string = '';
    numberTrack: number = 0;
    idBraiderAppoint: number = 0;
    idBraiderOwner: number = 0;
    idClientAppoint: number = 0;
    idStyleAppoint: number = 0;
    idSizeAppoint: number = 0;
    idLenghtAppoint: number = 0;
    desigStyle: string = '';
    titleSize: string = '';
    titleExtrat: string = '';
    clientFullName: string = '';
    serviceTitle: string = '';
    takeDownTitle: string = '';
    totalCostBraiding: number = 0;
}