// src/models/ExtratStyle.ts
export class ExtratStyle{
    idExtratStyle: number = 0;
    idStyle: number = 0;
    idExtrat: number = 0;
    idSize: number = 0;
    costExtra: number = 0.0;
    costTouchUpExtra: number = 0.0;
    costExtraSize: number = 0.0;
    costBusyExtra: number = 0.0;
    costHairDeductExtra: number = 0.0;
    //desigStyle: string = '';
    //titleSize: string = '';
    //titleExtrat: string = '';
}